moved patch_base_app and base_app from conftest.py to test/app/test_base.py
