+++ tests/conftest.py
@@ -2,9 +2,7 @@ from unittest.mock import patch
 
 import pytest
 from fakeredis import FakeRedis
-from pytest_mock import MockerFixture
 
-from corva.app.base import BaseApp
 from corva.app.scheduled import ScheduledApp
 from corva.app.stream import StreamApp
 from corva.constants import STREAM_EVENT_TYPE
@@ -45,22 +43,6 @@ def redis_adapter(patch_redis_adapter):
 @pytest.fixture(scope='function')
 def redis(redis_adapter):
     return RedisState(redis=redis_adapter)
-
-
-@pytest.fixture(scope='function')
-def patch_base_app(mocker: MockerFixture):
-    """Patches BaseApp
-
-    1. patches __abstractmethods__, so we can initialize BaseApp
-    """
-
-    mocker.patch.object(BaseApp, '__abstractmethods__', set())
-    yield
-
-
-@pytest.fixture(scope='function')
-def base_app(patch_base_app):
-    return BaseApp(app_key=APP_KEY, cache_url=CACHE_URL)
 
 
 @pytest.fixture(scope='function')
